// src/package/interface/package.interfaces.ts
 export interface PackageListParam {
    // day: number | null;
    // month: number | null;
    start: string;
    end: string | null;
    type: string| null;
    status: number | null;
    store_id: number | null;
    keyword_customer: string | null;
    keyword_card: string |  null;
    take: number | null;
    page: number | null;
  }
  